ml_worker/requirements.txt
celery
redis
torch
torchvision
Pillow
pymongo
python-dotenv
